+++ src/Ledger.jl
@@ -8,7 +8,7 @@ The current version of the ledger API.
 If the API version of JlrsLedger_jll and this version don't match, this module will fail to
 initialize.
 """
-const LEDGER_API_VERSION = 0x2
+const LEDGER_API_VERSION = 0x3
 
 const API_VERSION_FN = Ref{Ptr{Cvoid}}(C_NULL)
 const IS_BORROWED_SHARED = Ref{Ptr{Cvoid}}(C_NULL)
@@ -16,7 +16,6 @@ const IS_BORROWED_EXCLUSIVE = Ref{Ptr{Cvoid}}(C_NULL)
 const IS_BORROWED = Ref{Ptr{Cvoid}}(C_NULL)
 const BORROW_SHARED = Ref{Ptr{Cvoid}}(C_NULL)
 const BORROW_EXCLUSIVE = Ref{Ptr{Cvoid}}(C_NULL)
-const BORROW_SHARED_UNCHECKED = Ref{Ptr{Cvoid}}(C_NULL)
 const UNBORROW_SHARED = Ref{Ptr{Cvoid}}(C_NULL)
 const UNBORROW_EXCLUSIVE = Ref{Ptr{Cvoid}}(C_NULL)
 
@@ -25,7 +24,7 @@ export LedgerError, PoisonError, is_borrowed, is_borrowed_exclusive, is_borrowed
        unborrow_shared, LEDGER_API_VERSION
 
 """
-An exception that indicates the ledger was used incorrectly and has likely been corrupted.
+An exception that indicates the ledger was used incorrectly.
 """
 struct LedgerError <: Exception end
 
@@ -34,17 +33,15 @@ struct LedgerError <: Exception end
 
 Returns `true` if there is at least one active shared borrow of this data.
 """
-function is_borrowed_shared(@nospecialize data::Any)
+function is_borrowed_shared(data)
     if !ismutable(data)
         return false
     end
 
-    res = ccall(IS_BORROWED_SHARED[], UInt8, (Ptr{Cvoid},), Base.pointer_from_objref(data))
-    if res == 0x0
-        false
-    elseif res == 0x1
-        true
-    elseif res == 0x2
+    res = ccall(IS_BORROWED_SHARED[], Int32, (Ptr{Cvoid},), data)
+    if res >= 0
+        Bool(res)
+    else
         throw(LedgerError())
     end
 end
@@ -54,17 +51,15 @@ end
 
 Returns `true` if this data is exclusively borrowed.
 """
-function is_borrowed_exclusive(@nospecialize data::Any)
+function is_borrowed_exclusive(data)
     if !ismutable(data)
         return false
     end
 
-    res = ccall(IS_BORROWED_EXCLUSIVE[], UInt8, (Ptr{Cvoid},), Base.pointer_from_objref(data))
-    if res == 0x0
-        false
-    elseif res == 0x1
-        true
-    elseif res == 0x2
+    res = ccall(IS_BORROWED_EXCLUSIVE[], Int32, (Ptr{Cvoid},), data)
+    if res >= 0
+        Bool(res)
+    else
         throw(LedgerError())
     end
 end
@@ -76,17 +71,15 @@ end
 Returns `true` if this data is borrowed. Equivalent to
 `is_borrowed_shared(data) || is_borrowed_exclusive(data)`.
 """
-function is_borrowed(@nospecialize data::Any)
+function is_borrowed(data)
     if !ismutable(data)
         return false
     end
 
-    res = ccall(IS_BORROWED[], UInt8, (Ptr{Cvoid},), Base.pointer_from_objref(data))
-    if res == 0x0
-        false
-    elseif res == 0x1
-        true
-    elseif res == 0x2
+    res = ccall(IS_BORROWED[], Int32, (Ptr{Cvoid},), data)
+    if res >= 0
+        Bool(res)
+    else
         throw(LedgerError())
     end
 end
@@ -98,39 +91,15 @@ Marks the data as being borrowed if the data isn't exclusively borrowed. Returns
 success, `false` if the data is already borrowed exclusively. If `true` is returned you must call
 `unborrow_shared` when you're done using it.
 """
-function try_borrow_shared(@nospecialize data::Any)
+function try_borrow_shared(data)
     if !ismutable(data)
         return true
     end
 
-    res = ccall(BORROW_SHARED[], UInt8, (Ptr{Cvoid},), Base.pointer_from_objref(data))
-    if res == 0x0
-        false
-    elseif res == 0x1
-        true
-    elseif res == 0x2
-        throw(LedgerError())
-    end
-end
-
-
-"""
-    borrowed_shared_unchecked(data)
-
-Marks the data as being borrowed. Always returns `true`, you must call  `unborrow_shared` when
-you're done using it.
-"""
-function borrow_shared_unchecked(@nospecialize data::Any)
-    if !ismutable(data)
-        return true
-    end
-
-    res = ccall(BORROW_SHARED_UNCHECKED[], UInt8, (Ptr{Cvoid},), Base.pointer_from_objref(data))
-    if res == 0x0
-        false
-    elseif res == 0x1
-        true
-    elseif res == 0x2
+    res = ccall(BORROW_SHARED[], Int32, (Ptr{Cvoid},), data)
+    if res >= 0
+        Bool(res)
+    else
         throw(LedgerError())
     end
 end
@@ -142,17 +111,15 @@ Marks the data as being borrowed exclusively if the data isn't already borrowed.
 success, `false` if the data is already borrowed. If `true` is returned you must call
 `unborrow_exclusive` when you're done using it.
 """
-function try_borrow_exclusive(@nospecialize data::Any)
+function try_borrow_exclusive(data)
     if !ismutable(data)
         return false
     end
 
-    res = ccall(BORROW_EXCLUSIVE[], UInt8, (Ptr{Cvoid},), Base.pointer_from_objref(data))
-    if res == 0x0
-        false
-    elseif res == 0x1
-        true
-    elseif res == 0x2
+    res = ccall(BORROW_EXCLUSIVE[], Int32, (Ptr{Cvoid},), data)
+    if res >= 0
+        Bool(res)
+    else
         throw(LedgerError())
     end
 end
@@ -167,17 +134,15 @@ wasn't present in the ledger.
 Each successfull call to `try_borrow_shared` and `borrow_shared_unchecked` must have a matching
 call to this function.
 """
-function unborrow_shared(@nospecialize data::Any)
+function unborrow_shared(data)
     if !ismutable(data)
         return true
     end
 
-    res = ccall(UNBORROW_SHARED[], UInt8, (Ptr{Cvoid},), Base.pointer_from_objref(data))
-    if res == 0x0
-        false
-    elseif res == 0x1
-        true
-    elseif res == 0x2
+    res = ccall(UNBORROW_SHARED[], Int32, (Ptr{Cvoid},), data)
+    if res >= 0
+        Bool(res)
+    else
         throw(LedgerError())
     end
 end
@@ -191,17 +156,15 @@ ledger, a `LedgerError` is thrown if the data wasn't present in the ledger.
 
 Each successfull call to `try_borrow_exclusive` must have a matching call to this function.
 """
-function unborrow_exclusive(@nospecialize data::Any)
+function unborrow_exclusive(data)
     if !ismutable(data)
         return true
     end
 
-    res = ccall(UNBORROW_EXCLUSIVE[], UInt8, (Ptr{Cvoid},), Base.pointer_from_objref(data))
-    if res == 0x0
-        false
-    elseif res == 0x1
-        true
-    elseif res == 0x2
+    res = ccall(UNBORROW_EXCLUSIVE[], Int32, (Ptr{Cvoid},), data)
+    if res >= 0
+        Bool(res)
+    else
         throw(LedgerError())
     end
 end
@@ -223,7 +186,6 @@ function __init__()
     IS_BORROWED[] = Libdl.dlsym(libjlrs_ledger_handle, "jlrs_ledger_is_borrowed")
     BORROW_SHARED[] = Libdl.dlsym(libjlrs_ledger_handle, "jlrs_ledger_try_borrow_shared")
     BORROW_EXCLUSIVE[] = Libdl.dlsym(libjlrs_ledger_handle, "jlrs_ledger_try_borrow_exclusive")
-    BORROW_SHARED_UNCHECKED[] = Libdl.dlsym(libjlrs_ledger_handle, "jlrs_ledger_borrow_shared_unchecked")
     UNBORROW_SHARED[] = Libdl.dlsym(libjlrs_ledger_handle, "jlrs_ledger_unborrow_shared")
     UNBORROW_EXCLUSIVE[] = Libdl.dlsym(libjlrs_ledger_handle, "jlrs_ledger_unborrow_exclusive")
 end